requirements.txt
numpy
scikit-learn
seaborn
matplotlib
easygui
tqdm
Pillow
webcolors